slimmed down code
