+++ tasks/chat_tasks.py
@@ -149,21 +149,6 @@ _MODEL_TEMPERATURE_CONFIG = {
     "gpt-4.1-nano": {"supports_temperature": True, "min": 0.0, "max": 2.0},
 }
 
-# Database 
-# DB_DSN = os.getenv("POSTGRES_DSN_POOL") # e.g., Supabase -> Connection -> Get Pool URL
-# # DB_POOL_MIN_SIZE = 5  # <-- if i had more compute
-# # DB_POOL_MAX_SIZE = 20
-# DB_POOL_MIN_SIZE = 2
-# DB_POOL_MAX_SIZE = 5
-
-# Initialize Redis sync client for pub/sub
-REDIS_LABS_URL = (
-    "redis://default:"
-    + os.getenv("REDIS_PASSWORD")
-    + "@"
-    + os.getenv("REDIS_PUBLIC_ENDPOINT")
-)
-redis_sync = redis.Redis.from_url(REDIS_LABS_URL, decode_responses=True)
 
 # ——— Streaming Mode Configuration ————————————————————————————————————————————
 
@@ -200,20 +185,6 @@ def should_use_streaming(user_id: str, project_id: str = None) -> bool:
     return use_streaming
 
 # ——— Database and Metrics Variables ———————————————————————————————————————————————————————————
-
-# Persistent event loop per worker process
-# worker_loop = None
-
-# Connection pools
-# async_db_pool: Optional[asyncpg.Pool] = None                            # Asyncpg async database connection pool
-# redis_pool: Optional[aioredis.ConnectionPool] = None                    # Redis connection pool
-
-# db_pool = async_db_pool                                                 # 🙊 monkey patch
-
-# Track pool initialization to prevent race conditions
-_sync_pool_lock = threading.Lock()
-_async_pool_init_lock = asyncio.Lock()
-_redis_init_lock = asyncio.Lock()
 
 # Preformance and metrics
 performance_monitor = None  # Performance tracking instance
@@ -278,325 +249,6 @@ class PerformanceMetrics:
     total_time: float = 0
     tokens_per_second: float = 0
     cache_hit_rate: float = 0
-
-# ——— Async Event Loop & Worker Management ———————————————————————————————————————
-
-def get_worker_loop():
-    """Get or create the persistent worker event loop"""
-    global worker_loop
-    if worker_loop is None or worker_loop.is_closed():
-        worker_loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(worker_loop)
-    return worker_loop
-
-# def run_async_in_worker(coro):
-#     """Execute async code in the persistent worker loop"""
-#     loop = get_worker_loop()
-#     return loop.run_until_complete(coro)
-
-# async def init_worker_pools():
-#     """Initialize all async resources once per worker"""
-#     global db_pool, redis_pool
-    
-#     if not db_pool:
-#         db_pool = await asyncpg.create_pool(
-#             dsn=DB_DSN,
-#             min_size=2,
-#             max_size=10,
-#             command_timeout=30,
-#             statement_cache_size=0, 
-#             server_settings={'application_name': 'rag_worker'}
-#         )
-#         logger.info("✅ DB pool initialized")
-    
-#     if not redis_pool:
-#         redis_pool = aioredis.ConnectionPool.from_url(
-#             REDIS_LABS_URL,
-#             max_connections=20,
-#             decode_responses=True
-#         )
-#         logger.info("✅ Redis pool initialized")
-
-# async def cleanup_worker_pools():
-#     """Cleanup all resources"""
-#     global db_pool, redis_pool
-    
-#     if db_pool:
-#         await db_pool.close()
-#         db_pool = None
-    
-#     if redis_pool:
-#         await redis_pool.disconnect() 
-#         redis_pool = None
-    
-#     logger.info("✅ Worker pools cleaned up")
-
-# ——— Database Connection Management ———————————————————————————————————————
-
-# ==== Asynchronous DB pool =======
-
-# async def initialize_async_database_pool() -> asyncpg.Pool:
-#     """
-#     🔧 IMPROVED: Initialize async database connection pool with proper lifecycle management
-    
-#     Key improvements:
-#     - Proper connection lifecycle management
-#     - Race condition prevention
-#     - Better error handling and recovery
-#     - Pool health monitoring
-#     """
-#     global db_pool
-    
-#     # Prevent multiple simultaneous initializations
-#     async with _async_pool_init_lock:
-#         # Check if pool already exists and is healthy
-#         if db_pool and not db_pool.is_closing():
-#             try:
-#                 # Quick health check
-#                 async with db_pool.acquire(timeout=5) as conn:
-#                     await conn.fetchval('SELECT 1')
-#                 logger.info("✅ Existing database pool is healthy")
-#                 return db_pool
-#             except Exception as e:
-#                 logger.warning(f"⚠️ Existing pool unhealthy, recreating: {e}")
-#                 await _close_db_pool_safely()
-    
-#     if not DB_DSN:
-#         raise ValueError("❌ POSTGRES_DSN environment variable not set")
-    
-#     logger.info("🏊 Creating optimized asyncpg connection pool...")
-    
-#     try:
-#         # Create new pool with improved settings
-#         db_pool = await asyncpg.create_pool(
-#             dsn=DB_DSN,
-#             min_size=DB_POOL_MIN_SIZE,
-#             max_size=DB_POOL_MAX_SIZE,
-            
-#             # Connection timeouts
-#             command_timeout=30,          # Individual command timeout
-#             timeout=60,                  # Connection acquisition timeout
-            
-#             # Performance settings
-#             server_settings={
-#                 'application_name': 'streaming_chat_worker',
-#                 'tcp_keepalives_idle': '300',     # Keep connections alive
-#                 'tcp_keepalives_interval': '30',
-#                 'tcp_keepalives_count': '3'
-#             },
-            
-#             # Compatibility settings
-#             statement_cache_size=0,      # pgBouncer compatibility
-            
-#             # Connection lifecycle management
-#             setup=_setup_connection,     # Run on each new connection
-#             init=_init_connection,       # Run once per connection
-#         )
-        
-#         # Verify pool health with a test query
-#         async with db_pool.acquire() as conn:
-#             result = await conn.fetchval('SELECT version()')
-#             logger.info(f"✅ Database pool initialized successfully. Version: {result[:50]}...")
-            
-#         # Register cleanup on process exit
-#         weakref.finalize(db_pool, _cleanup_db_pool)
-        
-#         return db_pool
-        
-#     except Exception as e:
-#         logger.error(f"❌ Database pool initialization failed: {e}")
-#         db_pool = None
-#         raise
-
-async def _setup_connection(conn: asyncpg.Connection):
-    """Run setup commands on each new connection"""
-    try:
-        # Set connection-level settings for better performance
-        await conn.execute("SET timezone = 'UTC'")
-        await conn.execute("SET statement_timeout = '30s'")
-        await conn.execute("SET lock_timeout = '10s'")
-    except Exception as e:
-        logger.warning(f"⚠️ Connection setup failed: {e}")
-
-# async def _init_connection(conn: asyncpg.Connection):
-#     """Run initialization commands once per connection"""
-#     try:
-#         # Prepare frequently used statements for better performance
-#         await conn.prepare("SELECT 1")
-#     except Exception as e:
-#         logger.warning(f"⚠️ Connection init failed: {e}")
-
-# async def _close_db_pool_safely():
-#     """Safely close database pool with proper cleanup"""
-#     global db_pool
-    
-#     if db_pool and not db_pool.is_closing():
-#         try:
-#             logger.info("🔄 Closing database pool...")
-            
-#             # Give active connections time to complete
-#             await asyncio.wait_for(db_pool.close(), timeout=30)
-            
-#             # Wait for all connections to actually close
-#             await db_pool.wait_closed()
-            
-#             logger.info("✅ Database pool closed successfully")
-#         except asyncio.TimeoutError:
-#             logger.warning("⚠️ Database pool close timeout, forcing termination")
-#             db_pool.terminate()
-#         except Exception as e:
-#             logger.error(f"❌ Database pool close error: {e}")
-#             db_pool.terminate()
-#         finally:
-#             db_pool = None
-
-# def _cleanup_db_pool():
-#     """Cleanup function for weakref finalization"""
-#     # This runs during garbage collection, so avoid complex async operations
-#     logger.info("🧹 Database pool cleanup triggered")
-    
-# ==== Redis Connection =============
-
-# async def initialize_redis_pool() -> aioredis.ConnectionPool:
-#     """
-#     🔧 IMPROVED: Initialize Redis connection pool with proper error handling
-#     """
-#     global redis_pool
-    
-#     async with _redis_init_lock:
-#         # Check if pool already exists and is healthy
-#         if redis_pool:
-#             try:
-#                 # Quick health check
-#                 async with aioredis.Redis(connection_pool=redis_pool) as r:
-#                     await asyncio.wait_for(r.ping(), timeout=5)
-#                 logger.info("✅ Existing Redis pool is healthy")
-#                 return redis_pool
-#             except Exception as e:
-#                 logger.warning(f"⚠️ Existing Redis pool unhealthy, recreating: {e}")
-#                 await _close_redis_pool_safely()
-    
-#     try:
-#         logger.info("🔗 Creating Redis connection pool...")
-        
-#         redis_pool = aioredis.ConnectionPool.from_url(
-#             REDIS_LABS_URL,
-#             max_connections=20,
-#             decode_responses=True,
-#             socket_timeout=30,
-#             socket_connect_timeout=10,
-#             retry_on_timeout=True,
-#             health_check_interval=30,  # Check connection health every 30s
-#         )
-        
-#         # Test connection
-#         async with aioredis.Redis(connection_pool=redis_pool) as r:
-#             pong = await asyncio.wait_for(r.ping(), timeout=10)
-#             logger.info(f"✅ Redis pool initialized successfully. Ping: {pong}")
-        
-#         # Register cleanup
-#         weakref.finalize(redis_pool, _cleanup_redis_pool)
-        
-#         return redis_pool
-        
-#     except Exception as e:
-#         logger.error(f"❌ Redis pool initialization failed: {e}")
-#         redis_pool = None
-#         raise
-
-# async def _close_redis_pool_safely():
-#     """Safely close Redis pool"""
-#     global redis_pool
-    
-#     if redis_pool:
-#         try:
-#             logger.info("🔄 Closing Redis pool...")
-#             await redis_pool.disconnect()
-#             logger.info("✅ Redis pool closed successfully")
-#         except Exception as e:
-#             logger.error(f"❌ Redis pool close error: {e}")
-#         finally:
-#             redis_pool = None
-
-# def _cleanup_redis_pool():
-#     """Cleanup function for Redis pool"""
-#     logger.info("🧹 Redis pool cleanup triggered")
-
-# ==== Context managers (uses global pool imports) =================
-
-# @asynccontextmanager
-# async def get_db_connection():
-#     """Use global pool"""
-#     pool = get_global_async_db_pool()
-#     if not pool:
-#         await init_async_pools()
-#         pool = get_global_async_db_pool()
-    
-#     async with pool.acquire() as conn:
-#         yield conn
-
-# @asynccontextmanager  
-# async def get_redis_connection():
-#     """Use global Redis pool"""
-#     pool = get_global_redis_pool()
-#     if not pool:
-#         await init_async_pools()
-#         pool = get_global_redis_pool()
-        
-#     async with aioredis.Redis(connection_pool=pool) as redis:
-#         yield redis
-
-# @asynccontextmanager
-# async def get_db_connection():
-#     """
-#     🔧 IMPROVED: Context manager for database connections with better error handling
-#     """
-#     if not db_pool:
-#         await initialize_async_database_pool()
-    
-#     connection = None
-#     try:
-#         # Acquire connection with timeout
-#         connection = await asyncio.wait_for(
-#             db_pool.acquire(), 
-#             timeout=30
-#         )
-        
-#         # Verify connection is still good
-#         await connection.fetchval('SELECT 1')
-        
-#         yield connection
-        
-#     except asyncio.TimeoutError:
-#         logger.error("❌ Database connection acquisition timeout")
-#         raise
-#     except asyncpg.InterfaceError as e:
-#         logger.error(f"❌ Database interface error: {e}")
-#         # Try to reinitialize pool if connection is corrupted
-#         if "another operation is in progress" in str(e).lower():
-#             logger.warning("🔄 Detected corrupted connection, reinitializing pool...")
-#             await _close_db_pool_safely()
-#             await initialize_async_database_pool()
-#         raise
-#     except Exception as e:
-#         logger.error(f"❌ Database connection error: {e}")
-#         raise
-#     finally:
-#         if connection:
-#             try:
-#                 # Always release connection back to pool
-#                 await db_pool.release(connection)
-#             except Exception as e:
-#                 logger.warning(f"⚠️ Connection release error: {e}")
-
-# @asynccontextmanager
-# async def get_redis_connection():
-#     """Context manager for Redis connections"""
-#     if not redis_pool:
-#         await initialize_redis_pool()
-    
-#     async with aioredis.Redis(connection_pool=redis_pool) as redis:
-#         yield redis
 
 # ——— Retry & Circuit Break Management ———————————————————————————————————————
 
@@ -715,30 +367,6 @@ class StreamingChatManager:
             logger.error(f"❌ StreamingChatManager initialization failed: {e}")
             self._initialized = False
             raise
-
-    # async def _check_db_pool_health(self) -> bool:
-    #     """Check if database pool is healthy"""
-    #     try:
-    #         pool = get_global_async_db_pool()
-    #         if not pool:
-    #             return False
-                
-    #         async with pool.acquire(timeout=5) as conn:
-    #             await conn.fetchval('SELECT 1')
-    #         return True
-    #     except Exception as e:
-    #         logger.warning(f"⚠️ DB pool health check failed: {e}")
-    #         return False
-    
-    # async def _check_redis_pool_health(self) -> bool:
-    #     """Check if Redis pool is healthy"""
-    #     try:
-    #         async with get_redis_connection() as redis:
-    #             await redis.ping()
-    #         return True
-    #     except Exception as e:
-    #         logger.warning(f"⚠️ Redis pool health check failed: {e}")
-    #         return False
 
     async def process_streaming_query(
         self,
@@ -937,122 +565,7 @@ class StreamingChatManager:
             )
         
         logger.info(f"📝 Created assistant message placeholder: {assistant_id}")
-    
-    # async def _stream_rag_response(
-    #     self,
-    #     assistant_message_id: str,
-    #     chat_session_id: str,
-    #     query: str,
-    #     relevant_chunks: List[Dict],
-    #     chat_history: List[Dict],
-    #     llm_client: Any,
-    #     provider: str
-    # ) -> StreamingResponse:
-    #     """
-    #     Core streaming response generator with real-time processing
         
-    #     Generator function that:
-    #     - 1. Determines if the stream is still going
-    #     - 2. Extracts citations from the chunk "content"
-    #     - 3. Extracts highlights (pertinent info like dates or facts)
-    #     - 4. Broadcasts the chink to the UI via websocket
-    #     """
-
-    #     # # [DELETEME] Initialize universal streaming manager
-    #     # streaming_manager = UniversalStreamingManager()
-
-    #     # Build enhanced context with system instructions (from YAML)
-    #     context = await self._build_enhanced_context(
-    #         query, relevant_chunks, chat_history
-    #     )
-        
-    #     accumulated_content = ""
-    #     citations = []
-    #     highlights = []
-    #     last_broadcast = time.time()
-    #     last_db_update = time.time()
-    #     content_buffer = ""
-        
-    #     try:
-    #         logger.info(f"🌊 Starting {provider} streaming...")
-            
-    #         # 🔀 Need async for to iterate over the async generator from llm_client
-    #         async for raw_chunk in llm_client.stream_chat(context):
-                
-    #             # 🎯 NORMALIZE: Convert provider-specific format to text
-    #             chunk_text = self.normalizer.extract_text(raw_chunk, provider)
-                
-    #             # Skip empty chunks
-    #             if not chunk_text:
-    #                 continue
-                
-    #             # Check for completion
-    #             logger.info(f"🔍 RAW CHUNK: {raw_chunk}")
-    #             if self.normalizer.is_completion_chunk(raw_chunk, provider):
-    #                 logger.info("✅ Stream completion detected")
-    #                 break
-                
-    #             # Accumulate content
-    #             accumulated_content += chunk_text
-    #             content_buffer += chunk_text
-                
-    #             # 1) --- Extract citations using your existing processor
-    #             doc_citations, seen_citations = self.citation_processor.extract_document_citations_from_chunks(
-    #                 accumulated_content, relevant_chunks, getattr(self, '_seen_citations', set())
-    #             )
-    #             if doc_citations:
-    #                 citations.extend(doc_citations)
-    #                 self._seen_citations = seen_citations
-    #                 await self._broadcast_citations(chat_session_id, doc_citations)
-                
-    #             # 2) --- Extract highlights (your existing method)
-    #             new_highlights = self._extract_highlights_from_chunk(chunk_text)
-    #             if new_highlights:
-    #                 highlights.extend(new_highlights)
-    #                 await self._broadcast_highlights(chat_session_id, new_highlights)
-                
-    #             # 3) --- Real-time WebSocket broadcasting (every 50ms)
-    #             if time.time() - last_broadcast > STREAMING_CONFIG["chunk_broadcast_interval"]:
-    #                 await self._broadcast_content_chunk(
-    #                     chat_session_id, assistant_message_id, chunk_text
-    #                 )
-    #                 last_broadcast = time.time()
-                
-    #             # 4) --- Batched database updates (every 100ms)
-    #             if time.time() - last_db_update > STREAMING_CONFIG["db_batch_update_interval"]:
-    #                 if content_buffer:
-    #                     await self._batch_update_content(assistant_message_id, content_buffer)
-    #                     content_buffer = ""
-    #                     last_db_update = time.time()
-            
-    #         # Final buffer flush
-    #         if content_buffer:
-    #             await self._batch_update_content(assistant_message_id, content_buffer)
-            
-    #         # 🆕 Enrich citations with link previews (your existing method)
-    #         enriched_citations = await self._enrich_citations_with_previews(citations)
-            
-    #         logger.info(f"🏁 Final accum. content {accumulated_content}...")
-    #         logger.info(f"📊 Streaming complete: {len(accumulated_content)} chars, {len(enriched_citations)} citations, {len(highlights)} highlights")
-            
-    #         return StreamingResponse(
-    #             content=accumulated_content,
-    #             citations=enriched_citations,
-    #             highlights=highlights,
-    #             metadata={
-    #                 "provider": provider,  # 🆕 Track provider
-    #                 "model": getattr(llm_client, 'model_name', 'unknown'),
-    #                 "chunks_used": len(relevant_chunks),
-    #                 "total_tokens": len(accumulated_content.split())
-    #             },
-    #             is_complete=True
-    #         )
-            
-    #     except Exception as e:
-    #         logger.error(f"❌ {provider} streaming failed: {e}")
-    #         await self._update_message_status(assistant_message_id, "error", str(e))
-    #         raise
-    
     async def _stream_rag_response(
         self,
         assistant_message_id: str,
@@ -2035,108 +1548,3 @@ def restart_chat_session(user_id: str, project_id: str) -> str:
 
     logger.info(f"✅ Chat session restarted: {old_session_id} → {new_session_id}")
     return new_session_id
-
-# ——— Worker Lifecycle Management ————————————————————————————————————————————————
-
-@worker_init.connect
-def worker_init_handler(sender, **kwargs):
-    """Initialize worker with persistent event loop"""
-    logger.info("🚀 Initializing worker...")
-    
-    # Initialize the persistent event loop
-    loop = get_worker_loop()
-    
-    # Pre-initialize resources
-    loop.run_until_complete(init_worker_pools())
-    
-    logger.info("✅ Worker initialized with persistent event loop")
-
-@worker_shutdown.connect
-def worker_shutdown_handler(sender, **kwargs):
-    """Graceful worker shutdown"""
-    logger.info("🛑 Shutting down worker...")
-    
-    try:
-        # Cleanup in the persistent loop
-        loop = get_worker_loop()
-        if loop and not loop.is_closed():
-            loop.run_until_complete(cleanup_worker_pools())
-            loop.close()
-        
-        logger.info("✅ Worker shutdown complete")
-        
-    except Exception as e:
-        logger.error(f"❌ Worker shutdown error: {e}")
-
-# ——— Health Check Functions ————————————————————————————————————————————————
-
-# async def check_db_pool_health() -> bool:
-#     """Check if database pool is healthy"""
-#     if not db_pool or db_pool.is_closing():
-#         return False
-    
-#     try:
-#         async with asyncio.timeout(5):
-#             async with db_pool.acquire() as conn:
-#                 await conn.fetchval('SELECT 1')
-#         return True
-#     except Exception as e:
-#         logger.warning(f"⚠️ Database pool health check failed: {e}")
-#         return False
-
-# async def check_redis_pool_health() -> bool:
-#     """Check if Redis pool is healthy"""
-#     if not redis_pool:
-#         return False
-#     try:
-#         async with asyncio.timeout(5):
-#             async with aioredis.Redis(connection_pool=redis_pool) as r:
-#                 await r.ping()
-#         return True
-#     except Exception as e:
-#         logger.warning(f"⚠️ Redis pool health check failed: {e}")
-#         return False
-    
-# ——— [TO BE DELETED] Legacy Compatibility Stubs ————————————————————————————————————————————————
-
-# # 🔄 Maintain original callback handler for backward compatibility
-# def publish_token(chat_session_id: str, token: str):
-#     """🔄 Legacy token publishing - now redirects to WebSocket broadcasting"""
-#     # This is now handled by the streaming manager's WebSocket broadcasting
-#     pass
-
-
-# class StreamToClientHandler(BaseCallbackHandler):
-#     """🔄 Legacy callback handler - maintained for compatibility"""
-    
-#     def __init__(self, session_id: str):
-#         self.session_id = session_id
-
-#     def on_llm_new_token(self, token: str, **kwargs) -> None:
-#         # Legacy callback - now handled by streaming manager
-#         pass
-
-
-# def get_chat_llm(model_name: str = "gpt-4o-mini", temperature: float = 0.7, callback_manager=None):
-#     """🔄 Legacy LLM factory - deprecated in favor of enhanced factory"""
-#     logger.warning("⚠️ get_chat_llm is deprecated - use CitationAwareLLMFactory instead")
-    
-#     from langchain_openai import ChatOpenAI
-    
-#     cfg = _MODEL_TEMPERATURE_CONFIG.get(model_name, {"supports_temperature": True})
-#     llm_kwargs = {
-#         "api_key": OPENAI_API_KEY,
-#         "model": model_name,
-#         "streaming": False,
-#     }
-    
-#     if cfg.get("supports_temperature", False):
-#         lo, hi = cfg.get("min", 0.0), cfg.get("max", 2.0)
-#         safe_temp = max(lo, min(temperature, hi))
-#         llm_kwargs["temperature"] = safe_temp
-
-#     if callback_manager:
-#         llm_kwargs["streaming"] = True
-#         llm_kwargs["callback_manager"] = callback_manager
-
-#     return ChatOpenAI(**llm_kwargs)
